index.php Changes In the .header-right div, the old form has been completely replaced by a new container <div id="add-column-container"> which holds a single button, <button id="btn-add-new-column">.
index.php Summary of Changes
Removed Old Code: I have removed the now-obsolete block of code that tried to find and add a listener to a static #add-column-form.
New initAddColumnButton Function: This new, self-contained function handles the entire lifecycle:
It waits for a click on the + New Column button.
When clicked, it replaces the button with the form HTML.
It listens for the form to be submitted (Enter) or for the user to click away (blur) to revert the UI back to the button.
The button is now functional, but it's always visible. Our final step is to use CSS to hide it on all views except the Tasks view.

style.css Changes
Removed Old Rules: In Section 3, I removed the obsolete rules that targeted #add-column-form-container.
Added New Rules: I added new rules that target our new container, #add-column-container. One rule hides it by default, and the other uses the .view-tasks-active class on the <body> to show it.